requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-decouple
Pillow
drf-yasg
django-cors-headers